refactor: hook and field reorganization - standardize field names to PascalCase - relocate hooks based on type, field hooks in `/fields/hooks` and collection hooks in `/collections/hooks`
