+++ src/collections/hooks/checkSquareHook.ts
