+++ src/collections/hooks/snakeCaseUploadsHook.ts
