+++ src/components/Hyperlink/index.tsx
@@ -1,4 +1,4 @@
-import { LinkField } from '@/fields/link/types';
+import { LinkField } from '@/fields/Link/types';
 import { cn } from '@/utilities/cn';
 import Link from 'next/link';
